requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
Django==5.0.3
django-bootstrap==0.2.4
django-bootstrap4==24.1
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-js-asset==2.2.0
pillow==10.3.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
